Add files via upload
This code works on Simple Shift Cipher algorithm also known as the Caesar cipher. It operates by shifting each character in the plaintext by a fixed number of positions in the ASCII table to produce the ciphertext. This type of algorithm falls under the category of Symmetric Key Algorithms, where the same key (shift value) is used for both encryption and decryption.